Remove mentions of multifit and multi-epoch fitting.
This unambiguously makes galaxy model-fitting a coadd operation, but
simply mentions possibilities for how we constrain stellar motions;
I think there are several plausible options there, and it's premature
to select one.

I've kept the phrase "Multi-epoch object characterization" in a few
places as that's not inaccurate - we just won't be doing pixel-level
fitting to multiple epochs at once.