Remove mention of Monte Carlo sampling from galaxy model fitting.
This was originally driven by shear estimation needs, and we now plan
to do that a different way (which will eat up at least the compute
resources we had (optimistically) intended to accomplish this with).
That replacement plan still needs to be added here, as there's now no
mention of shapes for shear at all, but I'll make that another commit.

This also leaves the discussion of full-likelihood characterization a
little awkward, as it now only applies to photometric redshifts, and
those are already somewhat irregularly mentioned here - they're in the
Object table, but not many of the processing descriptions.  I'm not
going to try to fix that now, as how we'll do photo-zs is another bag
of cats.