Update discussions of coadds to reflect current plans.
This mostly removes a bunch of coadd flavors from what we "will
produce and not retain", instead relegating them to a list of things we
will deliver software+configuration for; for most of these, we have had
no need to use them internally, and it's not clear we ever will.

It's worth calling out Best-Seeing Coadds in particular, which were
actually on the list of coadds we'd retain before (and we have
requirements on that).  I don't think they're sufficiently
scientifically useful in their own right, and I think the responsible
thing to do is get the relevant requirements changed.

I've also added a mention (with cross-ref) of templates in DRP, and
walked back some of the premature assertions about how we would make
them, emphasizing that there is a lot that's still TBD.